backend/pinecone_upload/requirements.txt
# 環境變量管理 - 用於從 .env 檔案載入環境變數
python-dotenv>=1.0.0
# AI 嵌入和語意理解 - 提供多語言文本向量化和語義分析能力
cohere>=4.20
# 向量數據庫 - 用於高效存儲和檢索向量數據
pinecone
pinecone-client==3.0.3
# 數值計算庫 - 提供高效的數組操作和數學函數
numpy==1.24.3
# 進度條顯示 - 用於長時間運行任務的視覺化進度